Add more work with logging

+++ src/main/java/de/Tuuby/AgentSimulator/logging/ILogger.java
@@ -1,8 +1,8 @@
 package de.Tuuby.AgentSimulator.logging;
 
 public interface ILogger {
-    void log(String data, int worldAge);
-    String get(int worldAge);
+    void log(String data, long worldAge);
+    String get(long worldAge);
     void writeToFile();
     void writeToConsole();
     //void writeToWindow(int lines);
